# backend/api-document/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Metrade API",
    "description": "This API is used for managing users, products, orders, order items, carts, cart items, token.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "The name of the category"
          },
          "description": {
            "type": "string",
            "example": "Category for all electronic products",
            "description": "Description of the category"
          },
          "parent_id": {
            "type": "string",
            "example": "60d5f489f9e7c4a5d37fbe28",
            "description": "Reference to the parent category ID"
          },
          "ancestors": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "References to all ancestor categories"
            },
            "example": [
              "60d5f489f9e7c4a5d37fbe27",
              "60d5f489f9e7c4a5d37fbe26"
            ],
            "description": "Array of ancestor categories"
          },
          "children": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "References to direct child categories"
            },
            "example": [
              "60d5f489f9e7c4a5d37fbe29",
              "60d5f489f9e7c4a5d37fbe30"
            ],
            "description": "Array of direct child categories"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Indicates whether the category is active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-10-04T08:40:51.620Z",
            "description": "Category creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-10-04T09:00:51.620Z",
            "description": "Category update timestamp"
          }
        }
      },
...
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    }
  },
  "security": [
    {
      "accessTokenCookie": [],
      "refreshTokenCookie": []
    }
  ],
  "paths": {
...
    "/categories/main-category/main-relationship": {
      "get": {
        "tags": [
          "Category"
        ],
        "security": [],
        "summary": "Get main categories with hierarchy",
        "operationId": "categoryHierarchy",
        "responses": {
          "200": {
            "description": "List of categories with hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No main categories found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}